Alter exec_quit to use UINT8 type for the retcode otherwise QUIT -1 doesn't work! (-1 is reserved for the thread message to mean not to bail out.)

+++ src/mainstate.c
@@ -1993,7 +1993,7 @@ void exec_xproc(void) {
 ** the operating system. It defaults to 0 (EXIT_SUCCESS)
 */
 void exec_quit(void) {
-  int32 retcode;
+  uint8 retcode;
 #ifdef DEBUG
   if (basicvars.debug_flags.functions) fprintf(stderr, ">>> Entered function mainstate.c:exec_quit\n");
 #endif /* DEBUG */
